backend/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.34.0
starlette==0.49.1
python-multipart==0.0.22
sqlalchemy[asyncio]==2.0.37
aiosqlite==0.20.0
pydantic[email]==2.10.4
pydantic-settings==2.7.0
alembic==1.14.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
onnx==1.19.0
pyasn1==0.6.3
onnxruntime==1.20.1
numpy==2.1.3
pillow==12.1.1
torch==2.8.0
torchvision==0.23.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
langdetect==1.0.9
